+++ hydra-node/test/Hydra/Chain/Direct/Contract/CollectCom.hs
@@ -34,9 +34,20 @@ import Hydra.Chain.Direct.Tx (
   mkInitialOutput,
  )
 import qualified Hydra.Contract.Commit as Commit
+import Hydra.Contract.Error (toErrorCode)
+import Hydra.Contract.Head (
+  HeadError (
+    DatumNotFound,
+    IncorrectUtxoHash,
+    MissingCommits,
+    STNotSpent,
+    SignerIsNotAParticipant
+  ),
+ )
 import qualified Hydra.Contract.Head as Head
 import qualified Hydra.Contract.HeadState as Head
 import Hydra.Contract.HeadTokens (headPolicyId)
+import Hydra.Contract.Util (UtilError (MintingOrBurningIsForbidden))
 import qualified Hydra.Data.ContestationPeriod as OnChain
 import qualified Hydra.Data.Party as OnChain
 import Hydra.Ledger.Cardano (genAdaOnlyUTxO, genTxIn, genVerificationKey)
@@ -182,31 +193,31 @@ data CollectComMutation
 genCollectComMutation :: (Tx, UTxO) -> Gen SomeMutation
 genCollectComMutation (tx, _utxo) =
   oneof
-    [ SomeMutation (Just "incorrect utxo hash") MutateOpenUTxOHash . ChangeOutput 0 <$> mutateUTxOHash
-    , SomeMutation (Just "missing commits") MutateNumberOfParties <$> do
+    [ SomeMutation (Just $ toErrorCode IncorrectUtxoHash) MutateOpenUTxOHash . ChangeOutput 0 <$> mutateUTxOHash
+    , SomeMutation (Just $ toErrorCode MissingCommits) MutateNumberOfParties <$> do
         moreParties <- (: healthyOnChainParties) <$> arbitrary
         pure $
           Changes
             [ ChangeInputHeadDatum $ replaceParties moreParties healthyCollectComInitialDatum
             , ChangeOutput 0 $ mutatedPartiesHeadTxOut moreParties headTxOut
             ]
-    , SomeMutation (Just "ST not spent") MutateHeadId <$> do
+    , SomeMutation (Just $ toErrorCode STNotSpent) MutateHeadId <$> do
         illedHeadResolvedInput <-
           mkHeadOutput
             <$> pure testNetworkId
             <*> fmap headPolicyId (arbitrary `suchThat` (/= testSeedInput))
             <*> pure (toUTxOContext $ mkTxOutDatum healthyCollectComInitialDatum)
         return $ ChangeInput healthyHeadInput illedHeadResolvedInput (Just $ toScriptData Head.CollectCom)
-    , SomeMutation (Just "signer is not a participant") MutateRequiredSigner <$> do
+    , SomeMutation (Just $ toErrorCode SignerIsNotAParticipant) MutateRequiredSigner <$> do
         newSigner <- verificationKeyHash <$> genVerificationKey
         pure $ ChangeRequiredSigners [newSigner]
-    , SomeMutation (Just "datum not found") MutateCommitToInitial <$> do
+    , SomeMutation (Just $ toErrorCode DatumNotFound) MutateCommitToInitial <$> do
         -- we're satisfied with "datum not found" as the current version of the validator will consider
         -- the initial input as if it were a commit input, hence fetching the datum which is expected
         -- in a commit and complaining that it did not find it
         (txIn, HealthyCommit{cardanoKey}) <- elements $ Map.toList healthyCommits
         pure $ ChangeInput txIn (toUTxOContext $ mkInitialOutput testNetworkId testSeedInput cardanoKey) Nothing
-    , SomeMutation (Just "minting or burning is forbidden") MutateTokenMintingOrBurning
+    , SomeMutation (Just $ toErrorCode MintingOrBurningIsForbidden) MutateTokenMintingOrBurning
         <$> (changeMintedTokens tx =<< genMintedOrBurnedValue)
     ]
  where