+++ hydra-plutus/src/Hydra/Contract/Head.hs
@@ -14,6 +14,7 @@ module Hydra.Contract.Head (
   compiledValidator,
   validatorScript,
   validatorHash,
+  HeadError (..),
 ) where
 
 import PlutusTx.Prelude
@@ -24,6 +25,7 @@ import Hydra.Contract.HeadState (Input (..), Signature, SnapshotNumber, State (.
 import Hydra.Contract.Util (hasST, mustNotMintOrBurn, (===))
 import Hydra.Data.ContestationPeriod (ContestationPeriod, addContestationPeriod, milliseconds)
 import Hydra.Data.Party (Party (vkey))
+import Hydra.Prelude (Show)
 import Plutus.Extras (ValidatorType, scriptValidatorHash, wrapValidator)
 import Plutus.V1.Ledger.Time (fromMilliSeconds)
 import Plutus.V1.Ledger.Value (valueOf)
@@ -59,6 +61,7 @@ import PlutusTx (CompiledCode)
 import qualified PlutusTx
 import qualified PlutusTx.AssocMap as Map
 import qualified PlutusTx.Builtins as Builtins
+import Hydra.Contract.Error (ToErrorCode (..))
 
 type DatumType = State
 type RedeemerType = Input
@@ -86,7 +89,7 @@ headValidator oldState input ctx =
     (Closed{utxoHash, contestationDeadline}, Fanout{numberOfFanoutOutputs}) ->
       checkFanout utxoHash contestationDeadline numberOfFanoutOutputs ctx
     _ ->
-      traceError "invalid head state transition"
+      traceError "H01"
 
 -- | On-Chain verification for 'Abort' transition. It verifies that:
 --
@@ -107,7 +110,7 @@ checkAbort ctx@ScriptContext{scriptContextTxInfo = txInfo} headCurrencySymbol pa
     && mustReimburseCommittedUTxO
  where
   mustBurnAllHeadTokens =
-    traceIfFalse "burnt token number mismatch" $
+    traceIfFalse "H02" $
       burntTokens == length parties + 1
 
   minted = getValue $ txInfoMint txInfo
@@ -118,7 +121,7 @@ checkAbort ctx@ScriptContext{scriptContextTxInfo = txInfo} headCurrencySymbol pa
       Just tokenMap -> negate $ sum tokenMap
 
   mustReimburseCommittedUTxO =
-    traceIfFalse "reimbursed outputs dont match" $
+    traceIfFalse "H03" $
       hashOfCommittedUTxO == hashOfOutputs
 
   hashOfOutputs =
@@ -182,14 +185,14 @@ checkCollectCom ctx@ScriptContext{scriptContextTxInfo = txInfo} (contestationPer
     && mustNotChangeParameters
     && everyoneHasCommitted
     && mustBeSignedByParticipant ctx headId
-    && traceIfFalse "ST not spent" (hasST headId val)
+    && traceIfFalse "H04" (hasST headId val)
  where
   mustCollectUtxoHash =
-    traceIfFalse "incorrect utxo hash" $
+    traceIfFalse "H05" $
       utxoHash == hashPreSerializedCommits collectedCommits
 
   mustNotChangeParameters =
-    traceIfFalse "changed parameters" $
+    traceIfFalse "H06" $
       parties' == parties
         && contestationPeriod' == contestationPeriod
         && headId' == headId
@@ -205,14 +208,14 @@ checkCollectCom ctx@ScriptContext{scriptContextTxInfo = txInfo} (contestationPer
           , headId = hId
           } ->
           (p, h, cp, hId)
-      _ -> traceError "wrong state in output datum"
+      _ -> traceError "H07"
   headAddress = mkHeadAddress ctx
 
   val =
     maybe mempty (txOutValue . txInInfoResolved) $ findOwnInput ctx
 
   everyoneHasCommitted =
-    traceIfFalse "missing commits" $
+    traceIfFalse "H08" $
       nTotalCommits == length parties
 
   (collectedCommits, nTotalCommits) =
@@ -283,7 +286,7 @@ checkClose ctx parties initialUtxoHash sig cperiod headPolicyId =
     && mustNotChangeParameters
  where
   mustPreserveValue =
-    traceIfFalse "head value is not preserved" $
+    traceIfFalse "H09" $
       val === val'
 
   val' = txOutValue . head $ txInfoOutputs txInfo
@@ -291,7 +294,7 @@ checkClose ctx parties initialUtxoHash sig cperiod headPolicyId =
   val = maybe mempty (txOutValue . txInInfoResolved) $ findOwnInput ctx
 
   hasBoundedValidity =
-    traceIfFalse "hasBoundedValidity check failed" $
+    traceIfFalse "H10" $
       tMax - tMin <= cp
 
   (closedSnapshotNumber, closedUtxoHash, parties', closedContestationDeadline, headId', contesters') =
@@ -306,37 +309,37 @@ checkClose ctx parties initialUtxoHash sig cperiod headPolicyId =
           , headId
           , contesters
           } -> (snapshotNumber, utxoHash, p, contestationDeadline, headId, contesters)
-      _ -> traceError "wrong state in output datum"
+      _ -> traceError "H07"
 
   checkSnapshot
     | closedSnapshotNumber > 0 =
-      traceIfFalse "invalid snapshot signature" $
+      traceIfFalse "H11" $
         verifySnapshotSignature parties closedSnapshotNumber closedUtxoHash sig
     | otherwise =
-      traceIfFalse "closed with non-initial hash" $
+      traceIfFalse "H12" $
         closedUtxoHash == initialUtxoHash
 
   checkDeadline =
-    traceIfFalse "incorrect closed contestation deadline" $
+    traceIfFalse "H13" $
       closedContestationDeadline == makeContestationDeadline cperiod ctx
 
   cp = fromMilliSeconds (milliseconds cperiod)
 
   tMax = case ivTo $ txInfoValidRange txInfo of
     UpperBound (Finite t) _ -> t
-    _InfiniteBound -> traceError "infinite upper bound"
+    _InfiniteBound -> traceError "H14"
 
   tMin = case ivFrom $ txInfoValidRange txInfo of
     LowerBound (Finite t) _ -> t
-    _InfiniteBound -> traceError "infinite lower bound"
+    _InfiniteBound -> traceError "H15"
 
   mustNotChangeParameters =
-    traceIfFalse "changed parameters" $
+    traceIfFalse "H06" $
       headId' == headPolicyId
         && parties' == parties
 
   mustInitializeContesters =
-    traceIfFalse "contesters non-empty" $
+    traceIfFalse "H16" $
       null contesters'
 
   ScriptContext{scriptContextTxInfo = txInfo} = ctx
@@ -392,7 +395,7 @@ checkContest ctx contestationDeadline contestationPeriod parties closedSnapshotN
     && mustPreserveValue
  where
   mustPreserveValue =
-    traceIfFalse "head value is not preserved" $
+    traceIfFalse "H09" $
       val === val'
 
   val' = txOutValue . head $ txInfoOutputs txInfo
@@ -400,7 +403,7 @@ checkContest ctx contestationDeadline contestationPeriod parties closedSnapshotN
   val = maybe mempty (txOutValue . txInInfoResolved) $ findOwnInput ctx
 
   mustBeNewer =
-    traceIfFalse "too old snapshot" $
+    traceIfFalse "H17" $
       contestSnapshotNumber > closedSnapshotNumber
 
   mustBeMultiSigned =
@@ -409,11 +412,11 @@ checkContest ctx contestationDeadline contestationPeriod parties closedSnapshotN
   mustBeWithinContestationPeriod =
     case ivTo (txInfoValidRange txInfo) of
       UpperBound (Finite time) _ ->
-        traceIfFalse "upper bound beyond contestation deadline" $ time <= contestationDeadline
-      _ -> traceError "contest: no upper bound defined"
+        traceIfFalse "H18" $ time <= contestationDeadline
+      _ -> traceError "H19"
 
   mustNotChangeParameters =
-    traceIfFalse "changed parameters" $
+    traceIfFalse "H06" $
       parties' == parties
         && headId' == headId
         && contestationPeriod' == contestationPeriod
@@ -421,14 +424,14 @@ checkContest ctx contestationDeadline contestationPeriod parties closedSnapshotN
   mustPushDeadline =
     if length contesters' == length parties'
       then
-        traceIfFalse "must not push deadline" $
+        traceIfFalse "H20" $
           contestationDeadline' == contestationDeadline
       else
-        traceIfFalse "must push deadline" $
+        traceIfFalse "H21" $
           contestationDeadline' == addContestationPeriod contestationDeadline contestationPeriod
 
   mustUpdateContesters =
-    traceIfFalse "contester not included" $
+    traceIfFalse "H22" $
       contesters' == contester : contesters
 
   (contestSnapshotNumber, contestUtxoHash, parties', contestationDeadline', contestationPeriod', headId', contesters') =
@@ -444,17 +447,17 @@ checkContest ctx contestationDeadline contestationPeriod parties closedSnapshotN
           , headId = hid
           , contesters = cs
           } -> (snapshotNumber, utxoHash, p, dl, cp, hid, cs)
-      _ -> traceError "wrong state in output datum"
+      _ -> traceError "H07"
 
   ScriptContext{scriptContextTxInfo = txInfo} = ctx
 
   contester =
     case txInfoSignatories txInfo of
       [signer] -> signer
-      _ -> traceError "wrong number of signers"
+      _ -> traceError "H23"
 
   checkSignedParticipantContestOnlyOnce =
-    traceIfFalse "signer already contested" $
+    traceIfFalse "H24" $
       contester `notElem` contesters
 {-# INLINEABLE checkContest #-}
 
@@ -467,15 +470,15 @@ checkFanout ::
 checkFanout utxoHash contestationDeadline numberOfFanoutOutputs ScriptContext{scriptContextTxInfo = txInfo} =
   hasSameUTxOHash && afterContestationDeadline
  where
-  hasSameUTxOHash = traceIfFalse "fannedOutUtxoHash /= closedUtxoHash" $ fannedOutUtxoHash == utxoHash
+  hasSameUTxOHash = traceIfFalse "H25" $ fannedOutUtxoHash == utxoHash
   fannedOutUtxoHash = hashTxOuts $ take numberOfFanoutOutputs txInfoOutputs
   TxInfo{txInfoOutputs} = txInfo
 
   afterContestationDeadline =
     case ivFrom (txInfoValidRange txInfo) of
       LowerBound (Finite time) _ ->
-        traceIfFalse "lower bound before contestation deadline" $ time > contestationDeadline
-      _ -> traceError "fanout: no lower bound defined"
+        traceIfFalse "H26" $ time > contestationDeadline
+      _ -> traceError "H27"
 {-# INLINEABLE checkFanout #-}
 
 --------------------------------------------------------------------------------
@@ -498,14 +501,14 @@ makeContestationDeadline :: ContestationPeriod -> ScriptContext -> POSIXTime
 makeContestationDeadline cperiod ScriptContext{scriptContextTxInfo} =
   case ivTo (txInfoValidRange scriptContextTxInfo) of
     UpperBound (Finite time) _ -> addContestationPeriod time cperiod
-    _ -> traceError "close: no upper bound defined"
+    _ -> traceError "H28"
 {-# INLINEABLE makeContestationDeadline #-}
 
 mkHeadAddress :: ScriptContext -> Address
 mkHeadAddress ctx =
   let headInput =
         fromMaybe
-          (traceError "script not spending a head input?")
+          (traceError "H29")
           (findOwnInput ctx)
    in txOutAddress (txInInfoResolved headInput)
 {-# INLINEABLE mkHeadAddress #-}
@@ -517,12 +520,12 @@ mustBeSignedByParticipant ::
 mustBeSignedByParticipant ScriptContext{scriptContextTxInfo = txInfo} headCurrencySymbol =
   case getPubKeyHash <$> txInfoSignatories txInfo of
     [signer] ->
-      traceIfFalse "signer is not a participant" $
+      traceIfFalse "H30" $
         signer `elem` (unTokenName <$> participationTokens)
     [] ->
-      traceError "no signers"
+      traceError "H31"
     _ ->
-      traceError "too many signers"
+      traceError "H32"
  where
   participationTokens = loop (txInfoInputs txInfo)
   loop = \case
@@ -550,8 +553,8 @@ headOutputDatum ctx =
 findTxOutDatum :: TxInfo -> TxOut -> Datum
 findTxOutDatum txInfo o =
   case txOutDatum o of
-    NoOutputDatum -> traceError "no output datum"
-    OutputDatumHash dh -> fromMaybe (traceError "datum not found") $ findDatum dh txInfo
+    NoOutputDatum -> traceError "H33"
+    OutputDatumHash dh -> fromMaybe (traceError "H34") $ findDatum dh txInfo
     OutputDatum d -> d
 {-# INLINEABLE findTxOutDatum #-}
 
@@ -580,14 +583,14 @@ hasPT headCurrencySymbol txOut =
 
 verifySnapshotSignature :: [Party] -> SnapshotNumber -> BuiltinByteString -> [Signature] -> Bool
 verifySnapshotSignature parties snapshotNumber utxoHash sigs =
-  traceIfFalse "signature verification failed" $
+  traceIfFalse "H35" $
     length parties == length sigs
       && all (uncurry $ verifyPartySignature snapshotNumber utxoHash) (zip parties sigs)
 {-# INLINEABLE verifySnapshotSignature #-}
 
 verifyPartySignature :: SnapshotNumber -> BuiltinByteString -> Party -> Signature -> Bool
 verifyPartySignature snapshotNumber utxoHash party signed =
-  traceIfFalse "party signature verification failed" $
+  traceIfFalse "H36" $
     verifyEd25519Signature (vkey party) message signed
  where
   message =
@@ -616,3 +619,83 @@ validatorScript = getValidator $ mkValidatorScript compiledValidator
 
 validatorHash :: ValidatorHash
 validatorHash = scriptValidatorHash validatorScript
+
+-- * Errors
+
+data HeadError
+  = InvalidHeadStateTransition
+  | BurntTokenNumberMismatch
+  | ReimbursedOutputsDontMatch
+  | STNotSpent
+  | IncorrectUtxoHash
+  | ChangedParameters
+  | WrongStateInOutputDatum
+  | MissingCommits
+  | HeadValueIsNotPreserved
+  | HasBoundedValidityCheckFailed
+  | InvalidSnapshotSignature
+  | ClosedWithNonInitialHash
+  | IncorrectClosedContestationDeadline
+  | InfiniteUpperBound
+  | InfiniteLowerBound
+  | ContestersNonEmpty
+  | TooOldSnapshot
+  | UpperBoundBeyondContestationDeadline
+  | ContestNoUpperBoundDefined
+  | MustNotPushDeadline
+  | MustPushDeadline
+  | ContesterNotIncluded
+  | WrongNumberOfSigners
+  | SignerAlreadyContested
+  | FannedOutUtxoHashNotEqualToClosedUtxoHash
+  | LowerBoundBeforeContestationDeadline
+  | FanoutNoLowerBoundDefined
+  | CloseNoUpperBoundDefined
+  | ScriptNotSpendingAHeadInput
+  | SignerIsNotAParticipant
+  | NoSigners
+  | TooManySigners
+  | NoOutputDatumError
+  | DatumNotFound
+  | SignatureVerificationFailed
+  | PartySignatureVerificationFailed
+  deriving (Show)
+
+instance ToErrorCode HeadError where
+  toErrorCode = \case
+    InvalidHeadStateTransition -> "H01"
+    BurntTokenNumberMismatch -> "H02"
+    ReimbursedOutputsDontMatch -> "H03"
+    STNotSpent -> "H04"
+    IncorrectUtxoHash -> "H05"
+    ChangedParameters -> "H06"
+    WrongStateInOutputDatum -> "H07"
+    MissingCommits -> "H08"
+    HeadValueIsNotPreserved -> "H09"
+    HasBoundedValidityCheckFailed -> "H10"
+    InvalidSnapshotSignature -> "H11"
+    ClosedWithNonInitialHash -> "H12"
+    IncorrectClosedContestationDeadline -> "H13"
+    InfiniteUpperBound -> "H14"
+    InfiniteLowerBound -> "H15"
+    ContestersNonEmpty -> "H16"
+    TooOldSnapshot -> "H17"
+    UpperBoundBeyondContestationDeadline -> "H18"
+    ContestNoUpperBoundDefined -> "H19"
+    MustNotPushDeadline -> "H20"
+    MustPushDeadline -> "H21"
+    ContesterNotIncluded -> "H22"
+    WrongNumberOfSigners -> "H23"
+    SignerAlreadyContested -> "H24"
+    FannedOutUtxoHashNotEqualToClosedUtxoHash -> "H25"
+    LowerBoundBeforeContestationDeadline -> "H26"
+    FanoutNoLowerBoundDefined -> "H27"
+    CloseNoUpperBoundDefined -> "H28"
+    ScriptNotSpendingAHeadInput -> "H29"
+    SignerIsNotAParticipant -> "H30"
+    NoSigners -> "H31"
+    TooManySigners -> "H32"
+    NoOutputDatumError -> "H33"
+    DatumNotFound -> "H34"
+    SignatureVerificationFailed -> "H35"
+    PartySignatureVerificationFailed -> "H36"